DO NOT PULL THIS VERSION. DO NOT PULL.
this is writting screenshots to a permanent c path im using this for my submission. i will change it back. DO NOT PULL THIS PUSH, PULL THE PREVIOUS PUSH OR THE ONE AFTER.

im only pushing it because i fixed some minor bugs in it and i dont wanna lose all my fixes in case something happens